fix unit test for the get data from remote use case
